backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2
xhtml2pdf
pandas
